Part 1, separate swap and send, move duplicate code into separate files
