+++ src/components/CurrencyInputPanel/index.tsx
@@ -3,13 +3,14 @@ import React, { useState, useContext } from 'react'
 import styled, { ThemeContext } from 'styled-components'
 import '@reach/tooltip/styles.css'
 import { darken } from 'polished'
+import { Field } from '../../state/swap/actions'
 import { useTokenBalanceTreatingWETHasETH } from '../../state/wallet/hooks'
 
 import TokenLogo from '../TokenLogo'
 import DoubleLogo from '../DoubleLogo'
 import SearchModal from '../SearchModal'
 import { RowBetween } from '../Row'
-import { TYPE, Hover } from '../../theme'
+import { TYPE, CursorPointer } from '../../theme'
 import { Input as NumericalInput } from '../NumericalInput'
 import { ReactComponent as DropDown } from '../../assets/images/dropdown.svg'
 
@@ -120,8 +121,8 @@ const StyledBalanceMax = styled.button`
 
 interface CurrencyInputPanelProps {
   value: string
-  field: number
-  onUserInput: (field: number, val: string) => void
+  field: string
+  onUserInput: (field: string, val: string) => void
   onMax: () => void
   atMax: boolean
   label?: string
@@ -176,7 +177,7 @@ export default function CurrencyInputPanel({
                 {label}
               </TYPE.body>
               {account && (
-                <Hover>
+                <CursorPointer>
                   <TYPE.body
                     onClick={onMax}
                     color={theme.text2}
@@ -188,7 +189,7 @@ export default function CurrencyInputPanel({
                       ? 'Balance: ' + userTokenBalance?.toSignificant(6)
                       : ' -'}
                   </TYPE.body>
-                </Hover>
+                </CursorPointer>
               )}
             </RowBetween>
           </LabelRow>
@@ -249,7 +250,7 @@ export default function CurrencyInputPanel({
           showSendWithSwap={showSendWithSwap}
           hiddenToken={token?.address}
           otherSelectedTokenAddress={otherSelectedTokenAddress}
-          otherSelectedText={field === 0 ? 'Selected as output' : 'Selected as input'}
+          otherSelectedText={field === Field.INPUT ? 'Selected as output' : 'Selected as input'}
         />
       )}
     </InputPanel>
