+++ src/components/ExchangePage/styleds.ts
@@ -75,29 +75,6 @@ export const StyledNumerical = styled(NumericalInput)`
     color: ${({ theme }) => theme.text4};
   }
 `
-
-export const MaxButton = styled.button`
-  position: absolute;
-  right: 70px;
-  padding: 0.25rem 0.35rem;
-  background-color: ${({ theme }) => theme.primary5};
-  border: 1px solid ${({ theme }) => theme.primary5};
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  text-transform: uppercase;
-  cursor: pointer;
-  margin-right: 0.5rem;
-  color: ${({ theme }) => theme.primaryText1};
-  :hover {
-    border: 1px solid ${({ theme }) => theme.primary1};
-  }
-  :focus {
-    border: 1px solid ${({ theme }) => theme.primary1};
-    outline: none;
-  }
-`
-
 export const StyledBalanceMaxMini = styled.button<{ active?: boolean }>`
   height: 22px;
   width: 22px;
