+++ src/pages/App.tsx
@@ -9,7 +9,7 @@ import NavigationTabs from '../components/NavigationTabs'
 import Web3ReactManager from '../components/Web3ReactManager'
 
 import Popups from '../components/Popups'
-import { isAddress, getAllQueryParams } from '../utils'
+import { isAddress } from '../utils'
 
 import Swap from './Swap'
 import Send from './Send'
@@ -101,8 +101,6 @@ function GoogleAnalyticsReporter({ location: { pathname, search } }: RouteCompon
 }
 
 export default function App() {
-  const params = getAllQueryParams()
-
   return (
     <>
       <Suspense fallback={null}>
@@ -117,11 +115,9 @@ export default function App() {
               <Web3ReactManager>
                 <Body>
                   <NavigationTabs />
-                  {/* this Suspense is for route code-splitting */}
                   <Switch>
-                    <Route exact strict path="/" render={() => <Redirect to="/swap" />} />
-                    <Route exact strict path="/swap" component={() => <Swap params={params} />} />
-                    <Route exact strict path="/send" component={() => <Send params={params} />} />
+                    <Route exact strict path="/swap" component={Swap} />
+                    <Route exact strict path="/send" component={Send} />
                     <Route exact strict path="/find" component={() => <Find />} />
                     <Route exact strict path="/create" component={() => <Create />} />
                     <Route exact strict path="/pool" component={() => <Pool />} />
@@ -159,7 +155,7 @@ export default function App() {
                         }
                       }}
                     />
-                    <Redirect to="/" />
+                    <Redirect to="/swap" />
                   </Switch>
                 </Body>
               </Web3ReactManager>
